requirements.txt
asgiref==3.9.1
Django==5.2.3
django-allauth==65.10.0
django-extensions==4.1
django-unfold==0.62.0
django-widget-tweaks==1.5.0
psycopg==3.2.9
python-decouple==3.8
redis==6.2.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
